requirements.txt
jira>=3.5.0
google-genai>=0.2.0
anthropic[vertex]>=0.40.0
google-cloud-aiplatform>=1.38.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
PyYAML>=6.0
